% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gif_id.R
\name{gif_id}
\alias{gif_id}
\title{Get GIF by ID Endpoint}
\usage{
gif_id(id = NULL, api_key = "dc6zaTOxFJmzC")
}
\arguments{
\item{id}{The unique GIF identifier, or a character vector of IDs.}

\item{api_key}{- Giphy provides a default \code{api_key} value for public beta testing: \code{"dc6zaTOxFJmzC"}.  This is the default value.}
}
\description{
Get GIF metadata based on the unique id.
}
