% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gif_search.R
\name{gif_search}
\alias{gif_search}
\title{Search GIFs by key words throught giphy API}
\usage{
gif_search(query, limit = 10, offset = 0, rating = "g",
  img_format = c("fixed_height", "fixed_height_still",
  "fixed_height_downsampled", "fixed_width", "fixed_width_still",
  "fixed_width_downsampled", "fixed_height_small", "fixed_height_small_still",
  "fixed_width_small", "fixed_width_small_still", "downsized",
  "downsized_still", "downsized_large", "original", "original_still"))
}
\description{
Search GIFs by key words throught giphy API
}

