% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.giphy.R
\name{plot.giphy}
\alias{plot.giphy}
\title{Plot a GIF}
\usage{
\method{plot}{giphy}(x, n = 1, type = "original", ...)
}
\arguments{
\item{x}{a \code{giphy} object returned from one of the \code{giphy} commands.}

\item{n}{When the \code{giphy} contains multiple images, the index of the image requested.}

\item{type}{GIF size, either \code{original} or \code{downsized}.  Default is \code{original}.}

\item{...}{arguments to be passed to \code{magick::image_read}.}
}
\description{
Returns a random GIF from Giphy, which may be limited using tags.
}
\details{

}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
