% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_giphy.R
\name{plot_giphy}
\alias{plot_giphy}
\title{Plot a GIF}
\usage{
plot_giphy(image, n = 1, type = "original", action = TRUE)
}
\arguments{
\item{n}{When the \code{giphy} contains multiple images, the index of the image requested.}

\item{type}{GIF size, either \code{original} or \code{downsized}.  Default is \code{original}.}
}
\description{
Returns a random GIF from Giphy, which may be limited using tags.
}
\details{

}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
