% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.giphy.R
\name{print.giphy}
\alias{print.giphy}
\title{Print a giphy object}
\usage{
\method{print}{giphy}(image)
}
\arguments{
\item{image}{a \code{giphy} object.}
}
\description{
Print method for giphyR
}
\details{

}
\examples{

}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
