% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random}
\alias{random}
\title{Return a random GIF}
\usage{
random(tag = NULL, rating = NULL, sticker = FALSE,
  api_key = "dc6zaTOxFJmzC")
}
\arguments{
\item{tag}{a single, or multiple text strings.}

\item{rating}{- A rating for age appropriate gifs.  Options include \code{"y"}, \code{"g"}, \code{"pg"}, \code{"pg-13"} and \code{"r"}.}

\item{sticker}{Should a GIF be returned or an animated sticker (with transparent background)?  Default to \code{FALSE}.}

\item{api_key}{- Giphy provides a default \code{api_key} value for public beta testing: \code{"dc6zaTOxFJmzC"}.  This is the default value.}
}
\description{
Returns a random GIF from Giphy, which may be limited using tags.
}
\details{
Pulls a random GIF from the \code{giphy} API.  This may be limited through the use of tags.
    Be aware that failure to use the \code{rating} tag may result in gifs that are not safe for work use.
    The default API key used here is for testing purposes only.  More information on the use of the giphy API is available at \url{https://github.com/Giphy/GiphyAPI}
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
