% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{Giphy search}
\usage{
search(q, limit = 25, offset = 0, rating = NULL, lang = NULL,
  sticker = FALSE, api_key = "dc6zaTOxFJmzC")
}
\arguments{
\item{q}{Search query term or phrase}

\item{limit}{- The number of results to return, up to 100. The default is \code{limit=25}.}

\item{offset}{- An offset, to manage limited returns.  The default is \code{offset=0}, for subsequent searches, can increment \code{offset} by the size of \code{limit}}

\item{rating}{- A rating for age appropriate gifs.  Options include \code{"y"}, \code{"g"}, \code{"pg"}, \code{"pg-13"} and \code{"r"}.}

\item{lang}{- Regional content; see the allowed 2-letter ISO 639-1 country codes listed in the Details.}

\item{sticker}{Should a GIF be returned or an animated sticker (with transparent background)?  Default to \code{FALSE}.}

\item{api_key}{- Giphy provides a default \code{api_key} value for public beta testing: \code{"dc6zaTOxFJmzC"}.  This is the default value.}
}
\details{

}
\examples{

}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
