% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate Endpoint}
\usage{
translate(s, rating = NULL, lang = NULL, sticker = FALSE,
  api_key = "dc6zaTOxFJmzC")
}
\arguments{
\item{s}{- A term or phrase to translate into a GIF}

\item{rating}{- A rating for age appropriate gifs.  Options include \code{"y"}, \code{"g"}, \code{"pg"}, \code{"pg-13"} and \code{"r"}.}

\item{lang}{- Regional content; see the allowed 2-letter ISO 639-1 country codes listed in the Details.}

\item{sticker}{Should a GIF be returned or an animated sticker (with transparent background)?  Default to \code{FALSE}.}

\item{api_key}{- Giphy provides a default \code{api_key} value for public beta testing: \code{"dc6zaTOxFJmzC"}.  This is the default value.}
}
\description{
The translate API draws on search, but uses the Giphy
"special sauce" to handle translating from one vocabulary to another.
In this case, words and phrases to GIFs. Example implementations of
translate can be found in the Giphy Slack, Hipchat, Wire, or
Dasher integrations. Use a plus or url encode for phrases.
}
\details{
Tanslates a word or phrase to a GIF using the \code{giphy} API.
    Be aware that failure to use the \code{rating} tag may result in gifs that are not safe for work use.
    The default API key used here is for testing purposes only.  More information on the use of the giphy API is available at \url{https://github.com/Giphy/GiphyAPI}
    Language support is documented in the giphy API documentation for Language Support: \url{https://github.com/Giphy/GiphyAPI#language-support}.
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
