% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trending.R
\name{trending}
\alias{trending}
\title{Trending GIFs}
\usage{
trending(limit = 25, rating = NULL, sticker = FALSE,
  api_key = "dc6zaTOxFJmzC")
}
\arguments{
\item{limit}{- The number of results to return, up to 100. The default is \code{limit=25}.}

\item{rating}{- A rating for age appropriate gifs.  Options include \code{"y"}, \code{"g"}, \code{"pg"}, \code{"pg-13"} and \code{"r"}.}

\item{sticker}{Should a GIF be returned or an animated sticker (with transparent background)?  Default to \code{FALSE}.}

\item{api_key}{- Giphy provides a default \code{api_key} value for public beta testing: \code{"dc6zaTOxFJmzC"}.  This is the default value.}
}
\description{
Fetch GIFs currently trending online. Hand curated by the Giphy editorial team. The data returned mirrors the GIFs showcased on the Giphy homepage. Returns 25 results by default.
}
\details{

}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
