% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.giphy.R
\name{plot.giphy}
\alias{plot.giphy}
\title{Plot a GIF}
\usage{
\method{plot}{giphy}(x, n = 1, type = "original", ...)
}
\arguments{
\item{x}{a \code{giphy} object returned from one of the \code{giphy} commands.}

\item{n}{When the \code{giphy} contains multiple images, the index of the image requested.}

\item{type}{GIF size, either \code{original} or \code{downsized}.  Default is \code{original}.}

\item{...}{arguments to be passed to \code{magick::image_read}.}
}
\description{
Plots the giphy into your view-port.
}
\details{
Plotting a giphy requires installation of the \code{imagemagick} program.  When a \code{giphy}
    object is returned from one of the calls, the user may plot the first object returned (default \code{n = 1})
    or, may select one of the images returned.  The limit is generally 25, but more or fewer may be returned
    depending on the user parameters set.
    A user may plot the original image, or, possibly, a downscaled version of the GIF, if such an object exists.
    It is possible to pass other parameters to the function, but they are unused.


}
\examples{

gifs <- trending('horses', rating = 'g')
plot(gifs, n = 1)

# Plot a smaller version of the same gif:
plot(gifs, n = 2, type = 'downsized')
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
