% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trending.R
\name{trending}
\alias{trending}
\title{Trending GIFs}
\usage{
trending(limit = 25, rating = NULL, sticker = FALSE,
  api_key = "dc6zaTOxFJmzC")
}
\arguments{
\item{limit}{- The number of results to return, up to 100. The default is \code{limit=25}.}

\item{rating}{- A rating for age appropriate gifs.  Options include \code{"y"}, \code{"g"}, \code{"pg"}, \code{"pg-13"} and \code{"r"}.}

\item{sticker}{Should a GIF be returned or an animated sticker (with transparent background)?  Default to \code{FALSE}.}

\item{api_key}{- Giphy provides a default \code{api_key} value for public beta testing: \code{"dc6zaTOxFJmzC"}.  This is the default value.}
}
\description{
Fetch GIFs currently trending online. Hand curated by the Giphy editorial team. The data returned mirrors the GIFs showcased on the Giphy homepage. Returns 25 results by default.
}
\details{
Finds gifs that are currently trending through giphy's services.
    Be aware that failure to use the \code{rating} tag may result in gifs that are not safe for work use.
    The default API key used here is for testing purposes only.  More information on the use of the giphy API is available at \url{https://github.com/Giphy/GiphyAPI}
}
\examples{
gif <- trending(rating = 'pg', limit = 10)
plot(gif, n = 2)

# Generate a smaller gif with transparent background using the sticker flag:
gif <- trending(rating = 'pg', limit = 10, sticker = TRUE)
plot(gif, n = 2)

}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
